requirements.txt
certifi==2020.6.20
cffi==1.14.0
cycler==0.10.0
future==0.18.2
jellyfish==0.8.2
kiwisolver==1.2.0
matplotlib==3.2.2
mkl-fft==1.0.15
mkl-random==1.1.1
mkl-service==2.3.0
numpy==1.18.1
olefile==0.46
opencv-python==4.2.0.34
pandas==1.0.5
Pillow==7.1.2
pycparser==2.20
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
scipy==1.4.1
seaborn==0.10.1
Shapely==1.7.0
six==1.15.0
torch==1.1.0
torchvision==0.3.0
vtk==8.2.0
